requirements.txt
torchvision==0.17.0
torch>=2.1.0
numpy<2
matplotlib
pillow
argparse
scikit-image
tqdm
lime
ultralytics
shapely
ipykernel